MS-2: prompt-format fn lever + adaptive styles (public 47.57 -> 57.21)
Reworks the single-post EXFIL attack around the discovery that giving the
target the exact http.post(...) call syntax (fn style) makes it emit the
call with almost no analysis -> ~2.1x replay throughput -> +20% board score.
fn is now the default prompt_style and drives both target cells; best banked
public score 57.205 (up from 47.565).

- attack.py: pluggable EXFIL directive styles, gemma JSON-envelope styles,
  staged Harmony-forge multipost lever, and knobs (prompt_style, k_mode,
  multihop_style, hops_ask, prebuilt, gemma_forge).
- notebook/submission.ipynb: synced to attack.py.
- README.md / HANDOFF.md: final results, board GGUF finding, gemma analysis,
  v14/v15 regression diagnoses.
- dev/: throughput/gemma/GGUF probes, style-baking + watch/submit tooling;
  eval_real.py & overnight_optimize.py forced to bf16.

+++ dev/eval_real.py
@@ -53,7 +53,7 @@ def _shared_agent_factory(model: str):
 
         print(f"Loading {model} backend once (GB10)...", flush=True)
         t = time.time()
-        backend = build_gpt_oss_backend()
+        backend = build_gpt_oss_backend(torch_dtype="bfloat16")  # avoid broken mxfp4 triton path on GB10
         print(f"  loaded in {time.time() - t:.0f}s", flush=True)
         return lambda: GPTOSSAgent(backend=backend)
     if model in ("gemma", "gemma_4"):
